% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basal_binlevel.R
\docType{data}
\name{Basal_binlevel}
\alias{Basal_binlevel}
\title{Transcriptome location, bin-level read counts and size factors for 26 samples
A dataset containing the transcriptome location, read counts of 1000 bins from 26 samples,
and size factors estimated from the whole transcriptome to
normalize sequencing depth of each sample}
\format{
A list with 3 elements
\describe{
  \item{bins}{A dataframe of 1000 obs and 5 variables, containing the transcriptome location for 1000 bins of length 50bps}
  \item{bincounts}{A dataframe of 1000 obs and 26 variables, caontaining the read counts for ``bins"}
  \item{sf}{A numerical vector, containing the size factors of 26 samples estimated from the whole transcriptome}
  ...
}
}
\usage{
data(Basal_binlevel)
}
\description{
Transcriptome location, bin-level read counts and size factors for 26 samples
A dataset containing the transcriptome location, read counts of 1000 bins from 26 samples,
and size factors estimated from the whole transcriptome to
normalize sequencing depth of each sample
}
\keyword{datasets}
